requirements.txt
django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1
stringcase==1.2.0
Werkzeug==1.0.1
djangorestframework_simplejwt